+++ tests/testthat/test-rpanet.R
@@ -2,6 +2,9 @@ test_that("rpanet with default preference functions", {
   # sample PA networks
   set.seed(1234)
   nstep <- 1e3
+  my_weight_sampler <- function(n) {
+    rgamma(n, shape = 5, scale = 0.2)
+  }
   for (method in c("linear", "binary", "bag", "bagx")) {
     if (method == "linear" | method == "binary") {
       control <- rpa_control_preference(
@@ -13,9 +16,7 @@ test_that("rpanet with default preference functions", {
         rpa_control_scenario(
           alpha = 0.2, beta = 0.4, gamma = 0.2, xi = 0.1, rho = 0.1
         ) +
-        rpa_control_edgeweight(
-          distribution = rgamma, dparams = list(shape = 5, scale = 0.2)
-        )
+        rpa_control_edgeweight(sampler = my_weight_sampler)
     } else if (method == "bag") {
       control <- rpa_control_preference(
         ftype = "default",
@@ -36,19 +37,29 @@ test_that("rpanet with default preference functions", {
         rpa_control_scenario(
           alpha = 0.2, beta = 0.4, gamma = 0.2, xi = 0.1, rho = 0.1
         ) +
-        rpa_control_edgeweight(
-          distribution = rgamma, dparams = list(shape = 5, scale = 0.2)
-        )
+        rpa_control_edgeweight(sampler = my_weight_sampler)
     }
-    initial.network1 <- rpanet(1e3, directed = TRUE, control = control)
-    initial.network2 <- rpanet(1e3, directed = FALSE, control = control)
+    initial.network1 <- rpanet(1e3,
+      initial.network = list(
+        edgelist = matrix(1:2, nrow = 1),
+        directed = TRUE
+      ),
+      control = control
+    )
+    initial.network2 <- rpanet(1e3,
+      initial.network = list(
+        edgelist = matrix(1:2, nrow = 1),
+        directed = FALSE
+      ),
+      control = control
+    )
     net1 <- rpanet(
       control = control, nstep = nstep, initial.network = initial.network1,
-      directed = TRUE, method = method
+      method = method
     )
     net2 <- rpanet(
       control = control, nstep = nstep, initial.network = initial.network2,
-      directed = FALSE, method = method
+      method = method
     )
 
     # check node preference
@@ -109,17 +120,29 @@ test_that("rpanet with customized preference functions", {
         alpha = 0.2, beta = 0.4, gamma = 0.2, xi = 0.1, rho = 0.1
       ) +
       rpa_control_edgeweight(
-        distribution = rgamma, dparams = list(shape = 5, scale = 0.2)
+        sampler = function(n) rgamma(n, shape = 5, scale = 0.2)
       )
-    initial.network1 <- rpanet(1e3, directed = TRUE, control = control)
-    initial.network2 <- rpanet(1e3, directed = FALSE, control = control)
+    initial.network1 <- rpanet(1e3,
+      initial.network = list(
+        edgelist = matrix(1:2, nrow = 1),
+        directed = TRUE
+      ),
+      control = control
+    )
+    initial.network2 <- rpanet(1e3,
+      initial.network = list(
+        edgelist = matrix(1:2, nrow = 1),
+        directed = FALSE
+      ),
+      control = control
+    )
     net1 <- rpanet(
       control = control, nstep = nstep, initial.network = initial.network1,
-      directed = TRUE, method = method
+      method = method
     )
     net2 <- rpanet(
       control = control, nstep = nstep, initial.network = initial.network2,
-      directed = FALSE, method = method
+      method = method
     )
 
     # check node preference
@@ -167,23 +190,35 @@ test_that("rpanet initial network", {
   ) + rpa_control_scenario(
     alpha = 0.2, beta = 0.4, gamma = 0.2, xi = 0.1, rho = 0.1
   ) + rpa_control_edgeweight(
-    distribution = rgamma, dparams = list(shape = 5, scale = 0.2)
+    sampler = function(n) rgamma(n, shape = 5, scale = 0.2)
   ) + rpa_control_newedge(
-    distribution = rpois, dparams = list(lambda = 1), shift = 1
+    sampler = function(n) rpois(n, lambda = 2) + 1
   ) + rpa_control_reciprocal(
     group.prob = c(0.2, 0.4, 0.4),
     recip.prob = matrix(rep(0.5, 9), nrow = 3)
   )
 
-  netwk1 <- rpanet(1e3, directed = TRUE, control = ctr1)
+  netwk1 <- rpanet(1e3,
+    initial.network = list(
+      edgelist = matrix(1:2, nrow = 1),
+      directed = TRUE
+    ),
+    control = ctr1
+  )
   netwk2 <- rpanet(1e3,
     initial.network = netwk1,
-    directed = TRUE, control = netwk1$control
+    control = netwk1$control
   )
-  netwk3 <- rpanet(1e3, directed = FALSE, control = ctr1)
+  netwk3 <- rpanet(1e3,
+    initial.network = list(
+      edgelist = matrix(1:2, nrow = 1),
+      directed = FALSE
+    ),
+    control = ctr1
+  )
   netwk4 <- rpanet(1e3,
     initial.network = netwk3,
-    directed = FALSE, control = netwk1$control
+    control = netwk1$control
   )
 
   # check initial netwk
@@ -211,8 +246,20 @@ test_that("rpanet scenarios", {
   ctr <- rpa_control_scenario(
     alpha = 0.1, beta = 0.8, gamma = 0.1, beta.loop = FALSE
   )
-  netwk1 <- rpanet(1e4, control = ctr, directed = TRUE)
-  netwk2 <- rpanet(1e4, control = ctr, directed = FALSE)
+  netwk1 <- rpanet(1e4,
+    control = ctr,
+    initial.network = list(
+      edgelist = matrix(1:2, nrow = 1),
+      directed = TRUE
+    )
+  )
+  netwk2 <- rpanet(1e4,
+    control = ctr,
+    initial.network = list(
+      edgelist = matrix(1:2, nrow = 1),
+      directed = FALSE
+    )
+  )
 
   check_scenarios <- function(netwk) {
     alpha <- which(netwk$edge.attr$scenario == 1)
@@ -227,4 +274,24 @@ test_that("rpanet scenarios", {
 
   check_scenarios(netwk1)
   check_scenarios(netwk2)
+})
+
+test_that("rpanet node id", {
+  set.seed(12345)
+  ctr <- rpa_control_scenario(
+    alpha = 0.1, beta = 0.8, gamma = 0.1, beta.loop = FALSE
+  )
+  initial.network <- list(
+    edgelist = matrix(c(1010:1050, 3010:3050), ncol = 2), directed = TRUE
+  )
+  netwk1 <- rpanet(1e4,
+    control = ctr,
+    initial.network = initial.network
+  )
+  expect_true(
+    identical(
+      netwk1$edgelist[seq_len(nrow(initial.network$edgelist)), ],
+      initial.network$edgelist
+    )
+  )
 })